Update PR workflows and add comment feature
Updated the name of the workflow in `pull-request.yml` from "Pull Request Build and Test" to "Pull Request CI". Added a new workflow "Comment on the Pull Request" in `comment-on-pr.yml` that triggers upon completion of the "Pull Request CI" workflow. This new workflow runs on the latest Ubuntu version and comments on the PR only if the event that triggered the workflow is a pull request. The workflow includes steps to get the PR number, read it into GitHub environment variables, confirm the PR number for debugging, get the code coverage results file, and add a comment to the PR with the code coverage results. The comment is recreated each time the workflow runs.