Create API Project (#5)
* create project

* Directory.Packages.props added

* Directory.Build.props

* add editorconfig

* Updated build properties, added new projects and tests

Updated the condition for `PropertyGroup` and `ItemGroup` in `Directory.Build.props`. Added new packages to `Directory.Packages.props` and created a new project "tests" in `Neba.sln`. Updated `.editorconfig` with a new rule for `*.cs` files. Added a new `Directory.Build.props` file to the "tests" project with specific settings. Created a new `Neba.Api.UnitTests.csproj` project file and added a new test class `PlaceHolderTest` to the `Neba.Api.UnitTests` project.

* Updated Dockerfile with new COPY command

Updated the Dockerfile to include a new COPY command that copies the "Directory.Build.props" file from the source directory to the current directory in the Docker image. This change ensures consistent build settings across all projects in a solution by including common properties in the Docker image.

* Update Dockerfile and .dockerignore

* Update settings.json and Dockerfile ports

* Revert "Update settings.json and Dockerfile ports"

This reverts commit 25b9fd13a8ed31940b0158b1c77a783d40e7bb35.

* Revert "Update Dockerfile and .dockerignore"

This reverts commit 229f14677130414eea67a05839c9b747be9d5ec8.

* Update .gitignore

* "Set up GitHub Actions workflow for .NET project"

Implemented a GitHub Actions workflow for a .NET project that triggers on pull request events. The workflow runs on the latest Ubuntu and uses .NET version 8.0.X. It consists of two jobs: Build and Unit-Tests. The Build job checks out the code, sets up .NET, checks NuGet dependencies, restores project dependencies, and builds the project in Release configuration. The Unit-Tests job, which depends on the Build job, runs unit tests on the project and generates a coverage report. If triggered by a pull request event, the pull request number is saved and uploaded as an artifact. All artifacts are retained for 1 day.

* Update PR workflows and add comment feature

Updated the name of the workflow in `pull-request.yml` from "Pull Request Build and Test" to "Pull Request CI". Added a new workflow "Comment on the Pull Request" in `comment-on-pr.yml` that triggers upon completion of the "Pull Request CI" workflow. This new workflow runs on the latest Ubuntu version and comments on the PR only if the event that triggered the workflow is a pull request. The workflow includes steps to get the PR number, read it into GitHub environment variables, confirm the PR number for debugging, get the code coverage results file, and add a comment to the PR with the code coverage results. The comment is recreated each time the workflow runs.

* Update dotnet outdated command to exclude specific project in NEBA.sln

* Add coverlet.msbuild and ReportGenerator packages

* Add Calculator class and unit test for Add method