`Update GitHub Actions workflow for dev deployment`
Updated the `deploy-to-dev.yml` file to configure a GitHub Actions workflow for deploying an application to a development server. The workflow is triggered on a push event to the 'develop' and 'feature/*' branches. It consists of two jobs: `terraform` and `deploy-api`.

The `terraform` job sets up the environment variables for Terraform and runs several steps including code checkout, Terraform setup, initialization, validation, and application. Azure credentials are provided as secrets for the Terraform apply step.

The `deploy-api` job sets up the environment variables for the .NET application and runs steps including code checkout, .NET setup, dependency restoration, project build, project publish, Azure login, and API deployment to Azure. Azure credentials are provided as secrets for the Azure login step.