v0.1.0: Push API to Azure (#7)
* Create API Project (#5)

* create project

* Directory.Packages.props added

* Directory.Build.props

* add editorconfig

* Updated build properties, added new projects and tests

Updated the condition for `PropertyGroup` and `ItemGroup` in `Directory.Build.props`. Added new packages to `Directory.Packages.props` and created a new project "tests" in `Neba.sln`. Updated `.editorconfig` with a new rule for `*.cs` files. Added a new `Directory.Build.props` file to the "tests" project with specific settings. Created a new `Neba.Api.UnitTests.csproj` project file and added a new test class `PlaceHolderTest` to the `Neba.Api.UnitTests` project.

* Updated Dockerfile with new COPY command

Updated the Dockerfile to include a new COPY command that copies the "Directory.Build.props" file from the source directory to the current directory in the Docker image. This change ensures consistent build settings across all projects in a solution by including common properties in the Docker image.

* Update Dockerfile and .dockerignore

* Update settings.json and Dockerfile ports

* Revert "Update settings.json and Dockerfile ports"

This reverts commit 25b9fd13a8ed31940b0158b1c77a783d40e7bb35.

* Revert "Update Dockerfile and .dockerignore"

This reverts commit 229f14677130414eea67a05839c9b747be9d5ec8.

* Update .gitignore

* "Set up GitHub Actions workflow for .NET project"

Implemented a GitHub Actions workflow for a .NET project that triggers on pull request events. The workflow runs on the latest Ubuntu and uses .NET version 8.0.X. It consists of two jobs: Build and Unit-Tests. The Build job checks out the code, sets up .NET, checks NuGet dependencies, restores project dependencies, and builds the project in Release configuration. The Unit-Tests job, which depends on the Build job, runs unit tests on the project and generates a coverage report. If triggered by a pull request event, the pull request number is saved and uploaded as an artifact. All artifacts are retained for 1 day.

* Update PR workflows and add comment feature

Updated the name of the workflow in `pull-request.yml` from "Pull Request Build and Test" to "Pull Request CI". Added a new workflow "Comment on the Pull Request" in `comment-on-pr.yml` that triggers upon completion of the "Pull Request CI" workflow. This new workflow runs on the latest Ubuntu version and comments on the PR only if the event that triggered the workflow is a pull request. The workflow includes steps to get the PR number, read it into GitHub environment variables, confirm the PR number for debugging, get the code coverage results file, and add a comment to the PR with the code coverage results. The comment is recreated each time the workflow runs.

* Update dotnet outdated command to exclude specific project in NEBA.sln

* Add coverlet.msbuild and ReportGenerator packages

* Add Calculator class and unit test for Add method

* Push API (No Docker) (#6)

* remove dockerfile

* Update launchSettings.json

remove docker support

* Update launchSettings.json

* Added Azure Resource Manager provider and resources in Terraform

This commit introduces a new Azure Resource Manager (ARM) provider configuration in Terraform, enabling interaction with Azure resources through ARM APIs. It defines several variables with default values for resource group, app service plan, application insights, and API service.

The code creates a new Azure resource group, App Service Plan, Application Insights resource, and a Linux Web App with specified parameters. The Linux Web App configuration is set to always be off, HTTPS only, and includes the Application Insights instrumentation key in the app settings.

* `Update GitHub Actions workflow for dev deployment`

Updated the `deploy-to-dev.yml` file to configure a GitHub Actions workflow for deploying an application to a development server. The workflow is triggered on a push event to the 'develop' and 'feature/*' branches. It consists of two jobs: `terraform` and `deploy-api`.

The `terraform` job sets up the environment variables for Terraform and runs several steps including code checkout, Terraform setup, initialization, validation, and application. Azure credentials are provided as secrets for the Terraform apply step.

The `deploy-api` job sets up the environment variables for the .NET application and runs steps including code checkout, .NET setup, dependency restoration, project build, project publish, Azure login, and API deployment to Azure. Azure credentials are provided as secrets for the Azure login step.

* fix name

* update packages

* Update deploy-to-dev.yml

deploy depends on terraform

* add working directory

* move env variables up

* include subscription id

* Update Terraform setup and import resource group in deploy-to-dev workflow

* Fix typo in Terraform import command

* Fix typo in Terraform import command

* Update Azure backend configuration for Terraform

* Update environment variables and default values for dev deployment

Updated the `deploy-to-dev.yml` and `main.tf` files to remove the `storage_account_name` and `container_name` environment variables and add `resource_group_budget_cents` and `resource_group_budget_email`. The `terraform init` command was updated to reflect these changes. Default values for `resource_group_name`, `app_service_plan_name`, `app_insights_name`, and `api_service_name` in `main.tf` were updated to point to development resources. A new `azurerm_consumption_budget_resource_group` resource was added to set a budget for the resource group and send notifications when budget usage exceeds certain thresholds. The `azurerm_service_plan`, `azurerm_application_insights`, and `azurerm_linux_web_app` resources were updated to use the new default values.

* put back test as default

* add time

* Add budget action group for resource group consumption budget

* Update start_date format in azurerm_consumption_budget_resource_group

* Update time period format in azurerm_consumption_budget_resource_group

* Refactor start_date in azurerm_consumption_budget_resource_group

* Update time period in azurerm_consumption_budget_resource_group

* Update start date for consumption budget resource group

* Update Azure credentials format

* Add 'Development' environment variable to deploy-to-dev workflow

* Update deploy-to-dev.yml to remove unnecessary working directory

* Update Terraform and .NET setup versions, and specify runtime for API deployment

* Update resource group budget from cents to dollars

* Remove unused variable for resource group budget email

* Add dotnet publish command and display package path in Azure workflow

* Add system admin email to budget action group

* Update system admin email variable

* Add Terraform Plan and Wait for Approval steps, and update Publish step

* Fix dotnet publish command in deploy-to-dev workflow

* Refactor build and publish steps in deploy-to-dev workflow

* Update deploy-to-dev.yml and add working directory for Neba.Api

* Add dotnet publish step to deploy-to-dev workflow

* Update deploy-to-dev.yml to fix build and publish paths

* Update build and publish commands in deploy-to-dev.yml

* Add step to archive published files

* Update archive task name in deploy-to-dev workflow

* Remove unnecessary working directory reference

* Add manager email to deployment workflow and update variable descriptions

* Remove feature branch from deploy trigger

* Update Terraform plan output check