"Added MudBlazor package and configured its services"
Added the `MudBlazor` package version `6.17.0` to the `Directory.Packages.props` file and updated the `Neba.UI.csproj` file to include a reference to this package. The `_Imports.razor` and `Program.cs` files were updated to include the `MudBlazor` and `MudBlazor.Services` namespaces. In `Program.cs`, the `AddMudServices` method was called on the `Services` property of the `builder` object to add MudBlazor services to the application and the configuration for the Snackbar component was set.