Updated key vault and added new API URL variable
Updated the `nebamgmt-key-vault-name` value in the `development.tfvars` file and added a new variable `nebamgmt-api-url`. Also, declared the `nebamgmt-api-url` variable in the `main.tf` file with a description and a string type. A new `azurerm_key_vault_secret` resource named "nebamgmt-api-url-secret" has been added, which holds the value of the `nebamgmt-api-url` variable and is associated with the `azurerm_key_vault` resource named "nebamgmt-kv".