Added Application Insights telemetry to project
Added the `Microsoft.ApplicationInsights.AspNetCore` package version `2.22.0` to the `Directory.Packages.props` file to provide telemetry data tools for .NET Core applications. A reference to this package was also added to the `Neba.UI.csproj` file, enabling the project to use this package. Additionally, the `builder.Services.AddApplicationInsightsTelemetry();` line was added to the `Program.cs` file to enable the Application Insights telemetry service within the application.