Add manual trigger and Azure ARM env vars to GitHub Actions
This commit adds a `workflow_dispatch` event to the GitHub Actions workflow in the `deploy-to-dev.yml` file, allowing the workflow to be manually triggered from GitHub's UI. The `push` event is still present, so the workflow will also run automatically when changes are pushed to the 'develop', 'hotfix/*', and 'release/*' branches.

Additionally, the `env` section has been updated to set up environment variables for the Azure Resource Manager (ARM) using secrets stored in the GitHub repository. These variables include the Azure subscription ID, the Azure app client ID, and the Azure app client secret, which are necessary for deploying resources to Azure.