Add pull_request block to terraform-plan.yml
In this commit, a `pull_request` block has been added to the `terraform-plan.yml` file. This block defines actions to be taken when a pull request is opened, reopened, or synchronized, specifically for the 'develop' branch. The environment variables for Azure subscription ID, client ID, and client secret have not been modified.