`Update terraform-plan.yml with Azure env vars and dynamic env naming`
Added Azure related environment variables `ARM_SUBSCRIPTION_ID`, `ARM_CLIENT_ID`, `ARM_CLIENT_SECRET`, and `ARM_TENANT_ID` to `terraform-plan.yml`. These variables are set to the corresponding secrets stored in the GitHub repository. Also, updated the `terraform-plan` job to dynamically set the environment name based on the event that triggers the workflow. If the event is a pull request, the environment name is set to 'Development'. Otherwise, it uses the environment name provided in the event inputs.