v0.2.2: Fix Key Vault Permission (#15)
* Added new variables and resources to main.tf and development.tfvars

Added two new variables `application_name` in `main.tf` and `development.tfvars` to store and set the application name respectively. Introduced a new resource `azuread_application` named `nebamgmt-app` in `main.tf` to create an Azure AD application. Also added a new resource `azurerm_role_assignment` named `nebamgmt-app-kv-role` in `main.tf` to assign the "Key Vault Contributor" role to the Azure AD application with the scope set to the Key Vault resource `nebamgmt-kv`.

* "Updated attribute in azuread_application resource block"

""

"Changed the attribute from `name` to `display_name` in the `azuread_application` resource block in the `main.tf` file. This change ensures that the `application_name` variable is now used as the display name for the Azure AD application."

* Update application name and add new attribute in main.tf

Updated the `application_name` in the `development.tfvars` file from "NEBA Management (Dev)" to "nebamgmt-app-dev". Also, added a new attribute `content_type` with the value "text/url" to the `azurerm_key_vault_secret` resource named "nebamgmt-api-url-secret" in the `main.tf` file.

* Add manual trigger and Azure ARM env vars to GitHub Actions

This commit adds a `workflow_dispatch` event to the GitHub Actions workflow in the `deploy-to-dev.yml` file, allowing the workflow to be manually triggered from GitHub's UI. The `push` event is still present, so the workflow will also run automatically when changes are pushed to the 'develop', 'hotfix/*', and 'release/*' branches.

Additionally, the `env` section has been updated to set up environment variables for the Azure Resource Manager (ARM) using secrets stored in the GitHub repository. These variables include the Azure subscription ID, the Azure app client ID, and the Azure app client secret, which are necessary for deploying resources to Azure.

* Add pull_request block to terraform-plan.yml

In this commit, a `pull_request` block has been added to the `terraform-plan.yml` file. This block defines actions to be taken when a pull request is opened, reopened, or synchronized, specifically for the 'develop' branch. The environment variables for Azure subscription ID, client ID, and client secret have not been modified.

* `Update terraform-plan.yml with Azure env vars and dynamic env naming`

Added Azure related environment variables `ARM_SUBSCRIPTION_ID`, `ARM_CLIENT_ID`, `ARM_CLIENT_SECRET`, and `ARM_TENANT_ID` to `terraform-plan.yml`. These variables are set to the corresponding secrets stored in the GitHub repository. Also, updated the `terraform-plan` job to dynamically set the environment name based on the event that triggers the workflow. If the event is a pull request, the environment name is set to 'Development'. Otherwise, it uses the environment name provided in the event inputs.