Update deploy-to-dev.yml and .gitignore files
Updated the `deploy-to-dev.yml` file by removing the `workflow_dispatch:` line and adding environment variables for ARM authentication. This could potentially affect the manual triggering of the workflow. Also, `.gitignore` file has been updated to ignore `.DS_Store` files. The change related to `.idea/*` files in `.gitignore` is inconsequential as the line was removed and then added back in.