Add IQuery interface in Neba.Application.Messaging
This commit introduces a new interface, `IQuery<TResult>`, in the `IQuery.cs` file. This interface is part of the `Neba.Application.Messaging` namespace and inherits from `IRequest<ErrorOr<TResult>>` from the `MediatR` library. The `MediatR` library is used for handling requests and responses in a CQRS pattern. The `ErrorOr<TResult>` is a custom type that represents either an error or a successful result of type `TResult`.