Updated IQueryHandler and added new using directives
Modified the IQueryHandler interface to be a public interface with generic parameters and added constraints. Also, two new using directives, `using ErrorOr;` and `using MediatR;`, have been added at the top of the file. The namespace declaration has been moved down in the file. The IQueryHandler interface now extends IRequestHandler from the MediatR library, with TQuery as the request type and ErrorOr<TResult> as the response type.