Added Neba.Application reference and shared services
Added a reference to the `Neba.Application` project in the `Neba.Api` project file, allowing the use of classes and interfaces from `Neba.Application`. The `Neba.Application` namespace was also added to the `Program.cs` file for easier access to these classes and interfaces. A call to `builder.Services.AddSharedApplicationServices();` was added to `Program.cs` to add shared services to the dependency injection container. A new file `ApplicationDependencyInjection.cs` was created, containing a static class and method for adding services to the dependency injection container using the `MediatR` library and FluentValidation.