"Added new UTC time endpoint and problem details method"
Added `Neba.Application.Clock` to using directives, indicating usage of classes or interfaces from this namespace in the updated code. Introduced `AddProblemDetails()` method to the service container configuration to provide detailed problem responses in the API. A new endpoint mapping to the path `/utcNow` has been added, which uses an `IDateTimeProvider` to return the current UTC time, suggesting a new feature addition to the application.