Added docker-compose.yml to project and defined 'seq' service
Added a new solution item `docker-compose.yml` to the project. This file is used to define and run multi-container Docker applications. The `docker-compose.yml` file specifies the Docker Compose version as '3.4' and defines a service named 'seq'. This service uses the latest 'datalust/seq' image, is named 'neba.seq', and is set to always restart. Two environment variables are set, including 'ACCEPT_EULA' set to 'Y'. Ports '19630' and '19631' on the host are mapped to '80' and '5341' on the container respectively. A volume named 'neba.seq' is created and mapped to '/data' on the container.