Update Azure config and add feature flag management
Updated the `AddAzureAppConfiguration` method to accept a lambda function with an `options` object. This object is used to call the `Connect` method with the `connectionString`. Also, a new method `UseFeatureFlags` has been added to the `options` object. This method sets the `CacheExpirationInterval` to 10 seconds and calls the `Select` method with `KeyFilter.Any`. This indicates the use of Azure's feature management capabilities, specifically feature flags, with a cache refresh interval of 10 seconds.