Updated deployment workflow and Azure services configuration
Updated the `deploy-to-dev.yml` file to include two new required inputs for the `workflow_dispatch` event: `deployAPI` and `deployUI`. The conditions for running the 'Deploy API' and 'Deploy UI' jobs have been modified to only run if the previous job was successful and the corresponding input is `true`.

In the `InfrastructureConfiguration.cs` file, the way Azure App Configuration and Feature Management services are added has been changed. They are now added to a new `services` object and the Azure App Configuration is set up with a `Connect` and `UseFeatureFlags` method. The `UseFeatureFlags` method is configured with a `CacheExpirationInterval` of 10 seconds and a `KeyFilter` of `Any`. The `AddScopedFeatureManagement` method is now called on the `services` object.