Caching (#22)
* Update docker-compose.yml with new container names and add Redis cache service

* add caching feature

* add ICachedQuery

* add ICacheService

* query caching behavior

* cache serviced

* register caching

clean up old test feature flag.  need to think about how registering will be behing a feature flag

* handling feature flag for caching

* add rediscloud provider

will create script next.  also removing flag check.  we will always have a cache database, but will be lowest tier and flag off until we see a reason to include it and possibly increate the tier

* Add system-assigned identity to azurerm_app_configuration resource

* Add Redis cache configuration

* add source

* Add RedisCloud provider and update required version

* Update redis cache memory limit and throughput measurement value

* Add throughput measurement for Redis cache

* Add output for nebamgmt-redis-subscription-id and fix subscription_id value

* Update redis_cache_name variable in main.tf

* Fix subscription_id value in rediscloud_subscription_database resource

* Fix subscription ID in redis cache resource

* Remove unused Redis subscription

* remove redis infrastructure

* Add caching and clock functionality to the API

* Add feature flag for caching in UtcNowCache endpoint

* Update Program.cs

* Update InfrastructureConfiguration.cs

* Update InfrastructureConfiguration.cs

* Update InfrastructureConfiguration.cs

* Update InfrastructureConfiguration.cs

* Remove Azure login step from deploy-to-dev workflow

* Refactor configuration setup in Program.cs

* Refactor logging in InfrastructureConfiguration.cs

* Update appsettings.json

* Update appsettings.json

* Update InfrastructureConfiguration.cs

* Refactor app configuration and feature management

* Refactor feature management configuration and remove unused code

* Add Azure App Configuration service

* Refactor feature management configuration

* Update KeyVault configuration in UI and Terraform files

* Update Azure config and add feature flag management

Updated the `AddAzureAppConfiguration` method to accept a lambda function with an `options` object. This object is used to call the `Connect` method with the `connectionString`. Also, a new method `UseFeatureFlags` has been added to the `options` object. This method sets the `CacheExpirationInterval` to 10 seconds and calls the `Select` method with `KeyFilter.Any`. This indicates the use of Azure's feature management capabilities, specifically feature flags, with a cache refresh interval of 10 seconds.

* Updated deployment workflow and Azure services configuration

Updated the `deploy-to-dev.yml` file to include two new required inputs for the `workflow_dispatch` event: `deployAPI` and `deployUI`. The conditions for running the 'Deploy API' and 'Deploy UI' jobs have been modified to only run if the previous job was successful and the corresponding input is `true`.

In the `InfrastructureConfiguration.cs` file, the way Azure App Configuration and Feature Management services are added has been changed. They are now added to a new `services` object and the Azure App Configuration is set up with a `Connect` and `UseFeatureFlags` method. The `UseFeatureFlags` method is configured with a `CacheExpirationInterval` of 10 seconds and a `KeyFilter` of `Any`. The `AddScopedFeatureManagement` method is now called on the `services` object.

* Add `deployInfrastructure` param to control infra job

Added a new input parameter `deployInfrastructure` to the `workflow_dispatch` event in the `deploy-to-dev.yml` GitHub Actions workflow file. This boolean parameter, which defaults to `true`, determines whether the `infrastructure` job should be run. The `infrastructure` job now executes conditionally based on the value of `deployInfrastructure`.

* Update deployment conditions in `deploy-to-dev.yml`

Updated the conditional statements in the `deploy-to-dev.yml` file to check if the `deployInfrastructure`, `deployAPI`, and `deployUI` inputs are explicitly set to 'true'. This change provides a stricter condition for the deployment process, ensuring that these inputs are not just existing but specifically set to 'true'.

* Refactor AzureAppConfiguration and FeatureManagement setup

This commit includes several changes to the setup of AzureAppConfiguration and FeatureManagement. The `Microsoft.FeatureManagement` namespace was removed from the DEBUG and else conditions, and `Microsoft.Extensions.Configuration.AzureAppConfiguration` was added to the else condition. Method calls for adding AzureAppConfiguration and getting connection strings were updated to use the `builder` object instead of `services` and `configuration`. The method call for adding scoped FeatureManagement was also updated to use `builder.Services`.

* Create .gitattributes

* Remove StackExchangeRedis package and refactor internals visibility

Removed the "Microsoft.Extensions.Caching.StackExchangeRedis" package from both Directory.Packages.props and Neba.Infrastructure.csproj files. Also, the visibility of the internals in the Neba.Infrastructure.csproj file was refactored. The "InternalsVisibleTo" tags for "Neba.Infrastructure.UnitTests" and "DynamicProxyGenAssembly2" were removed and then added back, indicating a possible formatting change. The package references for "Microsoft.Azure.AppConfiguration.AspNetCore" and "Microsoft.FeatureManagement.AspNetCore" were temporarily removed and then added back in the Neba.Infrastructure.csproj file.

* Updated .gitattributes and modified Neba.sln

Updated the `.gitattributes` file to change the diff behavior for `.cs`, `.sln`, `.csproj`, `.jpg`, `.png`, and `.gif` files by uncommenting previously commented lines. In the `Neba.sln` file, added `.gitattributes` and `.gitignore` files to the "Solution Items" project and repositioned the `package-update.bat` file.

* update packages

* Remove RedisCloud access and secret keys from environment variables

* remove whitespace