Commit: Added health checks and updated logging in Neba.Api
This commit includes several changes primarily focused on adding health checks and updating logging in the Neba.Api project. The "AspNetCore.HealthChecks.UI.Client" package was added to both the Directory.Packages.props and Neba.Api.csproj files. The Serilog packages, including "Serilog.Sinks.Seq", were removed and re-added to the Neba.Api.csproj file, possibly due to code reordering or reformatting.

In the Program.cs file, the "HealthChecks.UI.Client" and "Microsoft.AspNetCore.Diagnostics.HealthChecks" namespaces were added, and health check mapping was implemented. The response writer for the health check varies depending on whether the code is in DEBUG mode or not.

Additionally, the "/weather" endpoint in the Program.cs file saw changes to its logging parameters. The parameter name was changed from "loggerFactory" to "factory", and the logger creation line was updated to use this new "factory" parameter.