Refactor Azure App Configuration and Key Vault setup
Refactored the `AddFeatureManagement` method in `InfrastructureDependencyInjection.cs` and `AddConfiguration` method in `InfrastructureConfiguration.cs` to modify how Azure App Configuration is added and connected. Instead of directly using `ManagedIdentityCredential` in the `Connect` method, it is now assigned to a variable `credential` and then used. Additionally, Key Vault is now configured with the same `credential` using the `ConfigureKeyVault` method. No changes were made to the use of feature flags and the addition of scoped feature management to the services.