Health Checks (#24)
* Add database connection string to configuration

* Update vault key reference for health check connection string

* Update main.tf

* Added SQL Server health checks and updated packages

Added the `AspNetCore.HealthChecks.SqlServer` package to the `Directory.Packages.props` and `Neba.Infrastructure.csproj` files for SQL Server health checks. The `AddCaching()` method in `InfrastructureDependencyInjection.cs` was replaced with `AddHealthChecks(configuration)`. Also, the `coverlet.collector`, `coverlet.msbuild`, `Microsoft.FeatureManagement`, and `Serilog.Sinks.Seq` packages were removed and re-added to the `Directory.Packages.props` file with the same versions, possibly due to formatting changes or mistakes.

* Commit: Added health checks and updated logging in Neba.Api

This commit includes several changes primarily focused on adding health checks and updating logging in the Neba.Api project. The "AspNetCore.HealthChecks.UI.Client" package was added to both the Directory.Packages.props and Neba.Api.csproj files. The Serilog packages, including "Serilog.Sinks.Seq", were removed and re-added to the Neba.Api.csproj file, possibly due to code reordering or reformatting.

In the Program.cs file, the "HealthChecks.UI.Client" and "Microsoft.AspNetCore.Diagnostics.HealthChecks" namespaces were added, and health check mapping was implemented. The response writer for the health check varies depending on whether the code is in DEBUG mode or not.

Additionally, the "/weather" endpoint in the Program.cs file saw changes to its logging parameters. The parameter name was changed from "loggerFactory" to "factory", and the logger creation line was updated to use this new "factory" parameter.

* Updated docker-compose and appsettings for new MSSQL service

Updated `docker-compose.yml` to include a new service for a Microsoft SQL Server database with the image `mcr.microsoft.com/mssql/server:2022-latest`. The `seq` service and `volumes` section were also updated. The `appsettings.Development.json` file was modified to include a new connection string for `HealthCheck` and to set the `Serilog:MinimumLevel:Default` setting to `Debug`.

* pass in managed identity to access keyvault

* use endpoint

* Refactor Azure App Configuration and Key Vault setup

Refactored the `AddFeatureManagement` method in `InfrastructureDependencyInjection.cs` and `AddConfiguration` method in `InfrastructureConfiguration.cs` to modify how Azure App Configuration is added and connected. Instead of directly using `ManagedIdentityCredential` in the `Connect` method, it is now assigned to a variable `credential` and then used. Additionally, Key Vault is now configured with the same `credential` using the `ConfigureKeyVault` method. No changes were made to the use of feature flags and the addition of scoped feature management to the services.

* Update app service plan and web app resources in Terraform files

Updated the `app_service_plan_sku_name` in `development.tfvars` from "F1" to "B1". In `main.tf`, made several changes to the `azurerm_linux_web_app` resources for "nebamgmt-api" and "nebamgmt-ui". For "nebamgmt-api", added an `application_stack` block with a `dotnet_version` of "8.0", a `remote_debugging_version` of "VS2022", a `health_check_path` of "/health", and a `health_check_eviction_time_in_min` of 5. For "nebamgmt-ui", changed the `always_on` variable from `var.api_always_on` to `var.ui_always_on`.

* fix seq logging

* Update VS Code settings for C# development

* add health check for keyvault

including secret for it to check

* Update deploy-to-dev.yml file

* Update ReportGenerator package version to 5.2.3

* Replace `deploy-to-dev.yml` with new `deploy.yml` workflow

This commit removes the old `deploy-to-dev.yml` file and introduces a new `deploy.yml` file. The new file contains a GitHub Actions workflow for deploying an application. The workflow is triggered on push events to the `develop`, `hotfix/*`, and `release/*` branches, and also allows manual triggering via the `workflow_dispatch` event. It defines three jobs: `terraform-apply`, `api`, and `ui`. The `terraform-apply` job sets up and runs Terraform to deploy infrastructure, while the `api` and `ui` jobs build, publish, and deploy an API and a UI application respectively. Both `api` and `ui` jobs depend on the successful completion of the `terraform-apply` job. The workflow uses several environment variables and secrets for deployment configuration, and utilizes GitHub Actions such as `actions/checkout@v4`, `hashicorp/setup-terraform@v3`, `actions/setup-dotnet@v4`, and `azure/login@v2`.